+++ src/com/dawkinstan/simplebalance/DbHelper.java
@@ -2,43 +2,53 @@ package com.dawkinstan.simplebalance;
 
 import android.content.Context;
 import android.database.sqlite.SQLiteDatabase;
+import android.database.sqlite.SQLiteDatabase.CursorFactory;
 import android.database.sqlite.SQLiteOpenHelper;
 import android.util.Log;
 
-public class DbHelper extends SQLiteOpenHelper {
+public class DbHelper extends SQLiteOpenHelper
+{
+	
+	private static final String LOGCAT = "DBLog";
+	
+	public static final String DATABASE_NAME = "info.db";
+	private static int DATABASE_VERSION = 1;
+	
+	public static final String TABLE_DATA = "data";
+	public static final String COLUMN_ID = "id";
+	public static final String COLUMN_TITLE = "title";
+	public static final String COLUMN_BEFORE_PURCHASE = "beforePurchase";
+	public static final String COLUMN_AFTER_PURCHASE = "afterPurchase";
+	public static final String COLUMN_WHERE_SPENT = "whereSpent";
+	public static final String COLUMN_AMOUNT_SPENT = "amountSpent";
+	public static final String COLUMN_TYPE = "type";
+	public static final String COLUMN_IS_DEPOSIT = "isDeposit";
+	
+	private static final String TABLE_CREATE = "" +
+			"CREATE TABLE "+TABLE_DATA+" ("+
+			COLUMN_ID + " INTEGER PRIMARY KEY AUTOINCREMENT, " +
+			COLUMN_AMOUNT_SPENT + " REAL, " +
+			COLUMN_BEFORE_PURCHASE + " REAL, " + 
+			COLUMN_AFTER_PURCHASE + " REAL, " +
+			COLUMN_WHERE_SPENT + " TEXT, " +
+			COLUMN_TYPE + " INTEGER" + "," +
+			COLUMN_TITLE + " TEXT" + ", " +
+			COLUMN_IS_DEPOSIT + " INTEGER" + ")";
+	
+	public DbHelper(Context context) {
+		super(context, DATABASE_NAME, null, DATABASE_VERSION);
+	}
 
-    private static final int DATABASE_VERSION = 1;
-    private static final String DATABASE_NAME = "simplebalance.db";
-    private static final String TABLE_BALANCE = "balancelog";
-    private static final String COLUMN_ID = "_id";
-    private static final String COLUMN_TRANSACTION_TYPE = "transactionType";
-    private static final String COLUMN_PURCHASE_TEXT = "purchaseText";
-    private static final String COLUMN_AMOUNT_TRANSACTED = "amountedTransacted";
-    
-    private static final String TABLE_CREATE = "create table "+TABLE_BALANCE+" ("
-    		+COLUMN_ID+" integer primary key autoincrement, " 
-    		+COLUMN_TRANSACTION_TYPE+"integer, "
-    		+COLUMN_PURCHASE_TEXT+" text not null, "
-    		+COLUMN_AMOUNT_TRANSACTED+" real"
-    		+");";
-
-    DbHelper(Context context) {
-        super(context, DATABASE_NAME, null, DATABASE_VERSION);
-    }
-
-    @Override
-    public void onCreate(SQLiteDatabase db) {
-        db.execSQL(TABLE_CREATE);
-    }
-
-
-	@Override
-	public void onUpgrade(SQLiteDatabase db, int oldVersion, int newVersion) {
-		// TODO Auto-generated method stub
-		Log.w(DbHelper.class.getName(),
-				"Upgrading database from version " + oldVersion + " to "
-		        + newVersion + ", which will destroy all old data");
-		db.execSQL("DROP TABLE IF EXISTS " + TABLE_BALANCE);
+	public void onCreate(SQLiteDatabase db)
+	{
+		db.execSQL(TABLE_CREATE);
+//		Log.(LOGCAT, "Table has been created");
+	}
+	
+	public void onUpgrade(SQLiteDatabase db, int oldVersion, int newVersion)
+	{
+		db.execSQL("DROP TABLE IF EXISTS " + TABLE_DATA);
 		onCreate(db);
 	}
+	
 }