+++ CHANGELOG.md
@@ -6,3 +6,8 @@
 * 01/21/13
  - addition/subtraction methods
  - set initial value
+* 05/03/13
+ - Added database to get info from new textboxes
+ - Added actionbar
+* 05/04/13
+ - Added log activity to view database